% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{activate_menu_theme}
\alias{activate_menu_theme}
\title{Activate RSCodeIO Menu Theme}
\usage{
activate_menu_theme(backup = TRUE, silent = FALSE)
}
\arguments{
\item{backup}{Logical. If TRUE, create backup of original files. Default is TRUE.}

\item{silent}{Logical. If TRUE, suppress messages. Default is FALSE.}
}
\value{
Logical. TRUE if activation was successful, FALSE otherwise.
}
\description{
Activates RSCodeIO styling in RStudio's file menus by replacing QSS files.
Not supported on Mac or RStudio Server.
}
\examples{
\dontrun{
activate_menu_theme()
activate_menu_theme(backup = FALSE)
}
}
