% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_theme_paths}
\alias{get_all_theme_paths}
\title{Get All Available Theme Paths}
\usage{
get_all_theme_paths()
}
\value{
A named list containing paths to theme files
}
\description{
Returns a named list with paths to all available theme files.
}
\examples{
\dontrun{
get_all_theme_paths()
}
}
