% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stylesheets_location}
\alias{get_stylesheets_location}
\title{Get Stylesheets Location}
\usage{
get_stylesheets_location()
}
\value{
A character string with the path to the stylesheets directory
}
\description{
Locates the directory containing QSS stylesheets within the package.
This function is not supported on Mac systems as QSS stylesheets are not used there.
}
\examples{
\dontrun{
get_stylesheets_location()
}
}
