% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{get_theme_status}
\alias{get_theme_status}
\title{Get Theme Installation Status}
\usage{
get_theme_status()
}
\value{
A named list with installation status information.
}
\description{
Returns detailed information about the current RSCodeIO theme installation status.
}
\examples{
\dontrun{
get_theme_status()
}
}
