% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gnome_theme_dark}
\alias{gnome_theme_dark}
\title{Get Gnome Dark Theme Path}
\usage{
gnome_theme_dark()
}
\value{
A character string with the path to the Gnome dark theme file
}
\description{
Returns the full path to the Gnome dark theme QSS file.
}
\examples{
\dontrun{
gnome_theme_dark()
}
}
