% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gnome_theme_dark_backup}
\alias{gnome_theme_dark_backup}
\title{Get Gnome Dark Theme Backup Path}
\usage{
gnome_theme_dark_backup()
}
\value{
A character string with the path to the Gnome dark theme backup file
}
\description{
Returns the full path to the Gnome dark theme backup QSS file.
}
\examples{
\dontrun{
gnome_theme_dark_backup()
}
}
