% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{host_os_is_mac}
\alias{host_os_is_mac}
\title{Check if Host OS is Mac}
\usage{
host_os_is_mac()
}
\value{
A logical value: TRUE if running on Mac, FALSE otherwise
}
\description{
Determines if the current operating system is macOS (Darwin).
}
\examples{
host_os_is_mac()
}
