% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_rstudio_server}
\alias{is_rstudio_server}
\title{Check if Running RStudio Server}
\usage{
is_rstudio_server()
}
\value{
A logical value: TRUE if running RStudio Server, FALSE otherwise
}
\description{
Determines if the current RStudio session is running in server mode.
}
\examples{
\dontrun{
is_rstudio_server()
}
}
