% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{print_theme_status}
\alias{print_theme_status}
\title{Print Theme Status}
\usage{
print_theme_status()
}
\value{
Nothing (called for side effects).
}
\description{
Prints a human-readable summary of the RSCodeIO theme installation status.
}
\examples{
\dontrun{
print_theme_status()
}
}
