% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{windows_theme_dark}
\alias{windows_theme_dark}
\title{Get Windows Dark Theme Path}
\usage{
windows_theme_dark()
}
\value{
A character string with the path to the Windows dark theme file
}
\description{
Returns the full path to the Windows dark theme QSS file.
}
\examples{
\dontrun{
windows_theme_dark()
}
}
