% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{windows_theme_dark_backup}
\alias{windows_theme_dark_backup}
\title{Get Windows Dark Theme Backup Path}
\usage{
windows_theme_dark_backup()
}
\value{
A character string with the path to the Windows dark theme backup file
}
\description{
Returns the full path to the Windows dark theme backup QSS file.
}
\examples{
\dontrun{
windows_theme_dark_backup()
}
}
