% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscodeio.R
\name{is_menu_theme_active}
\alias{is_menu_theme_active}
\title{Check if Menu Theme is Active}
\usage{
is_menu_theme_active()
}
\value{
Logical. TRUE if menu theme is active, FALSE otherwise.
}
\description{
Determines if the RSCodeIO menu theme is currently active by checking
for the presence of backup files.
}
\examples{
\dontrun{
is_menu_theme_active()
}
}
